feat: move command
